+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] - 2020-03-08
+
+- Refactor Track to use `Tone.Channel` instead of `Tone.PanVol`
+- Add Track `mute` and `solo`
+- Update `propTypes`
+
 ## [0.3.2] - 2020-03-03
 
 - Add Create React Template for Reactronica